rename file for pathlists
