+++ internal/impl/shell_test.go
@@ -13,7 +13,6 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"go.jetpack.io/devbox/internal/impl/envpath"
 	"go.jetpack.io/devbox/internal/shellgen"
 )
 
@@ -102,33 +101,6 @@ Generated shellrc != shellrc.golden (-shellrc.golden +shellrc):
 If the new shellrc is correct, you can update the golden file with:
 
 	go test -run "^%s$" -update`), diff, t.Name())
-			}
-		})
-	}
-}
-
-func TestCleanEnvPath(t *testing.T) {
-	tests := []struct {
-		name    string
-		inPath  string
-		outPath string
-	}{
-		{
-			name:    "NoEmptyPaths",
-			inPath:  "/usr/local/bin::",
-			outPath: "/usr/local/bin",
-		},
-		{
-			name:    "NoRelativePaths",
-			inPath:  "/usr/local/bin:/usr/bin:../test:/bin:/usr/sbin:/sbin:.:..",
-			outPath: "/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin",
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := envpath.JoinPathLists(test.inPath)
-			if got != test.outPath {
-				t.Errorf("Got incorrect cleaned PATH.\ngot:  %s\nwant: %s", got, test.outPath)
 			}
 		})
 	}
